lana_shop/requirements.txt
args==0.1.0
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.5.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
# diploma-frontend @ file:///Users/lana/PycharmProjects/Python_django_diploma_dpo/diploma-frontend/dist/diploma-frontend-0.2.tar.gz#sha256=c77c11799ca193e73d9cef8ac7744ae9f102bd4611210f4e809162964a94bf3a
Django==4.1.7
django-filter==22.1
django-import-export==3.0.2
django-js-asset==2.0.0
django-mptt==0.14.0
django-property-filter==1.1.2
django-rosetta==0.9.8
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
loguru==0.6.0
MarkupPy==1.14
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
Pillow==9.4.0
polib==1.1.1
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
self==2020.12.3
sqlparse==0.4.3
tablib==3.3.0
telebot==0.0.5
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
win32-setctime==1.1.0
xlrd==2.0.1
xlwt==1.3.0
